refactor(over window): move `build_changes` to `OverPartition` for better modularity (#18846)
